requirements.txt
asyncpg
fastapi
SQLAlchemy
uvicorn
redis
pycryptodome
python-multipart